feat: removed BaseProps
