+++ components/BaseDataDisplayProps/package.json
@@ -27,9 +27,7 @@
     "build": "rollup -c ../../rollup.config.mjs",
     "clean": "rimraf dist tsconfig.tsbuildinfo"
   },
-  "dependencies": {
-    "@gemeente-denhaag/baseprops": "workspace:*"
-  },
+  "dependencies": {},
   "peerDependencies": {
     "react": "^17.0.0 || ^18.0.0"
   },
