+++ components/BaseDataDisplayProps/src/index.tsx
@@ -1,7 +1,30 @@
 import React from 'react';
-import BaseProps, { BaseChildrenProps, BaseClassesProps } from '@gemeente-denhaag/baseprops';
 
-export default interface BaseDataDisplayProps extends BaseProps {
+export default interface BaseDataDisplayProps {
+  /**
+   * The contents of the component.
+   */
+  children?: React.ReactNode;
+
+  /**
+   * Override or extend the styles applied to the component.
+   */
+  classes?: Record<string, unknown>;
+
+  /**
+   * The tab-index applied to the root element of the component.
+   */
+  tabIndex?: number;
+
+  /**
+   * The id attribute is used to specify a unique id for an HTML element.
+   */
+  id?: string;
+
+  /**
+   * Extend the styles of the component by adding new classes.
+   */
+  className?: string;
   /**
    * The component used for the root node.
    * Either a string to use a HTML element or a component.
@@ -10,7 +33,11 @@ export default interface BaseDataDisplayProps extends BaseProps {
   component?: React.ElementType;
 }
 
-export interface BaseDataDisplayChildrenProps extends BaseChildrenProps {
+export interface BaseDataDisplayChildrenProps {
+  /**
+   * The contents of the component.
+   */
+  children?: React.ReactNode;
   /**
    * The component used for the root node.
    * Either a string to use a HTML element or a component.
@@ -19,7 +46,11 @@ export interface BaseDataDisplayChildrenProps extends BaseChildrenProps {
   component?: React.ElementType;
 }
 
-export interface BaseDataDisplayClassesProps extends BaseClassesProps {
+export interface BaseDataDisplayClassesProps {
+  /**
+   * Override or extend the styles applied to the component.
+   */
+  classes?: Record<string, unknown>;
   /**
    * The component used for the root node.
    * Either a string to use a HTML element or a component.
