+++ components/Button/package.json
@@ -29,9 +29,6 @@
   },
   "bugs": "https://github.com/nl-design-system/denhaag/issues",
   "dependencies": {
-    "@gemeente-denhaag/baseprops": "workspace:*"
-  },
-  "devDependencies": {
     "@gemeente-denhaag/icons": "workspace:*"
   },
   "peerDependencies": {
