+++ components/Divider/package.json
@@ -30,13 +30,10 @@
   },
   "dependencies": {
     "@gemeente-denhaag/basedatadisplayprops": "workspace:*",
-    "@gemeente-denhaag/baseprops": "workspace:*"
+    "@gemeente-denhaag/icons": "workspace:*"
   },
   "peerDependencies": {
     "react": "^17.0.0 || ^18.0.0"
-  },
-  "devDependencies": {
-    "@gemeente-denhaag/icons": "workspace:*"
   },
   "gitHead": "dcf72a9b79266c1ebede35aff4a02dd9121a980f"
 }
