+++ components/ResponsiveContent/package.json
@@ -24,11 +24,9 @@
   "files": [
     "dist"
   ],
-  "dependencies": {
-    "@gemeente-denhaag/baseprops": "workspace:*"
-  },
   "scripts": {
     "build": "rollup -c ../../rollup.config.mjs",
     "clean": "rimraf dist tsconfig.tsbuildinfo"
-  }
+  },
+  "dependencies": {}
 }
