+++ components/ResponsiveContent/src/index.tsx
@@ -1,11 +1,19 @@
 import React from 'react';
-import BaseProps from '@gemeente-denhaag/baseprops';
 import clsx from 'clsx';
 import './index.scss';
 
-export interface ResponsiveContentProps extends Omit<BaseProps, 'tabIndex'> {}
+export interface ResponsiveContentProps {
+  /**
+   * The contents of the component.
+   */
+  children?: React.ReactNode;
+  /**
+   * Extend the styles of the component by adding new classes.
+   */
+  className?: string;
+}
 
-export const ResponsiveContent: React.FC<BaseProps> = (props: ResponsiveContentProps) => {
+export const ResponsiveContent = (props: ResponsiveContentProps) => {
   const classNames = clsx('denhaag-responsive-content', props.className);
   return <div className={classNames}>{props.children}</div>;
 };
