+++ packages/components-react/package.json
@@ -31,7 +31,6 @@
     "@gemeente-denhaag/action": "workspace:*",
     "@gemeente-denhaag/alert": "workspace:*",
     "@gemeente-denhaag/badge-counter": "workspace:*",
-    "@gemeente-denhaag/baseprops": "workspace:*",
     "@gemeente-denhaag/button": "workspace:*",
     "@gemeente-denhaag/card": "workspace:*",
     "@gemeente-denhaag/checkbox": "workspace:*",
